\name{observed.data.G}
\alias{observed.data.G}
\docType{data}
\title{table of observed data}
\description{Data for use with examples for a main effect genetic exposure model}
\usage{data(observed.data.G)}
\format{
  A data frame with 5000 observations on the following 5 variables.
  \describe{
    \item{\code{id}}{scenario ID}
    \item{\code{phenotype}}{phenotype}
    \item{\code{genotype}}{genotypes}
    \item{\code{allele.A}}{A allele of the SNP}
    \item{\code{allele.B}}{B allele of the SNP}
  }
}
\examples{data(observed.data.G)}