\name{true.data.G}
\alias{true.data.G}
\docType{data}
\title{True data table}
\description{Data for use with examples for a main effect genetic exposure model}
\usage{data(true.data.G)}
\format{
  A data frame with 5000 observations on the following 5 variables.
  \describe{
    \item{\code{id}}{secneario ID}
    \item{\code{phenotype}}{phenotypes}
    \item{\code{genotype}}{genotypes}
    \item{\code{allele.A}}{A alleles}
    \item{\code{allele.B}}{B alleles}
  }
}
\examples{data(true.data.G)}

